[Feat] : 탈퇴 사유 받기 api 작성

+++ src/main/java/com/potato/balbambalbam/data/entity/Withdrawal.java
@@ -20,7 +20,7 @@ public class Withdrawal {
     private long userId;
 
     @Column(name = "code", nullable = false)
-    private int code;
+    private long code;
 
     @Column(name = "details")
     private String details;
